pyproject.toml
[project]
name = "chatbot-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "openai>=1.107.1",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pypdf>=4.3.1",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "openai-agents>=0.3.0",
    "mcp>=1.14.0",
    "SQLAlchemy>=2.0.35",
    "psycopg[binary]>=3.2.1",
]
